feat: reorganize Settings tabs for cleaner UX
Split the overloaded Recording tab into focused sections:
- New "Hotkeys" tab for all keyboard shortcut configuration
- New "About" tab with version info, update check, and copyright
- Move auto-unload and spoken feedback to Advanced tab
- Permissions section only shown when access is missing
- Extract HotkeyRecorderView into its own file
