+++ TypeWhisper/Views/GeneralSettingsView.swift
@@ -146,23 +146,6 @@ struct GeneralSettingsView: View {
                 }
             }
 
-            #if !APPSTORE
-            Section(String(localized: "Updates")) {
-                HStack {
-                    let version = Bundle.main.object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String ?? "?"
-                    let build = Bundle.main.object(forInfoDictionaryKey: "CFBundleVersion") as? String ?? "?"
-                    Text("Version \(version) (\(build))")
-                        .foregroundStyle(.secondary)
-
-                    Spacer()
-
-                    Button(String(localized: "Check for Updates...")) {
-                        UpdateChecker.shared?.checkForUpdates()
-                    }
-                    .disabled(UpdateChecker.shared?.canCheckForUpdates() != true)
-                }
-            }
-            #endif
         }
         .formStyle(.grouped)
         .padding()
